clean up _config.ts so it's literally only config info. everything else is now a plugin
