+++ _stable-urls.json
@@ -285,6 +285,7 @@
   "/posts/20260720-hey-yall-i-fucked-up/index.html",
   "/posts/20260809-having-to-constantly/index.html",
   "/posts/20260809-okay-posts-are-final/index.html",
+  "/posts/20260810-i-m-at-the-dmv-ama/index.html",
   "/posts/20260810-site-update-changelo/index.html",
   "/search/index.html",
   "/tags/-/index.html",
