+++ src/kitchen-sink.vto
@@ -21,18 +21,16 @@ index: false
     {
       reblogs: [
         {
-          data: {
-            description: "test",
-            url: "#",
-            title: "Test",
-            author: {
-              username: "ewie",
-              displayName: "Evie Finch",
-              avatar: "https://cdn.ewie.online/ewie-pfp.png",
-            },
-            date: new Date(),
-            tags: ["test", "testing"],
+          description: "test",
+          url: "#",
+          title: "Test",
+          author: {
+            username: "ewie",
+            displayName: "Evie Finch",
+            avatar: "https://cdn.ewie.online/ewie-pfp.png",
           },
+          date: new Date(),
+          tags: ["test", "testing"],
           content: `I am an AI hater. This is considered rude, but I do not care, because I am a hater.
           To speak politely about AI, you put disclaimers before criticism: of course I’m not against it entirely; perhaps in a few years when; maybe for other purposes, but. You are supposed to debate how and when it should be used. You are supposed to take for granted that it must be useful somewhere, to someone, for something, eventually. People who are rich and smart and respected are saying so, and it would be arrogant to disagree with such people.
           But I am a hater, which is a kind of integrity. It means I am willing to disagree with anyone, even if it is rude. “But I only use it to–” “Actually if you just—” “The new models–” “I was making fun–” Stop. You’re embarrassing yourself. I am embarrassed for you.
